+++ tests/archives/density_archive_test.py
@@ -1,0 +1,7 @@
+"""Tests for the DensityArchive."""
+
+from ribs.archives import DensityArchive
+
+# TODO: Test the buffer after adding solutions.
+# TODO: Test density values.
+# TODO: Check density dtype.