+++ HISTORY.md
@@ -20,7 +20,8 @@
 - Drop Python 3.8 support and remove pinned requirements {{pr}`497`)
 - **Backwards-incompatible:** BanditScheduler: Add emitter_pool and active attr;
   remove emitters attr ({pr}`494`)
-- Add DensityRanker for density descent search ({pr}`483`, {pr}`504`)
+- Add DensityArchive and DensityRanker for Density Descent Search ({pr}`483`,
+  {pr}`504`, {pr}`487`)
 - Add NoveltyRanker for novelty search ({pr}`477`)
 - Add proximity_archive_plot for visualizing ProximityArchive ({pr}`476`,
   {pr}`480`, {pr}`523`)